adding wifimanager and coin config

+++ btc_display/network.h
@@ -5,7 +5,7 @@
 #include <string>
 
 struct CoinData {
-  int price;
+  float price;
   float price_change_percentage;
   std::string symbol;
 };
